requirements.txt
annotated-types==0.7.0
anyio==4.8.0
autopep8==2.3.2
certifi==2024.12.14
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
distlib==0.3.9
fastapi==0.115.6
filelock==3.16.1
flake8==7.1.1
flake8-chart==0.1.5
gunicorn==23.0.0
h11==0.14.0
hacking==7.0.0
identify==2.6.5
idna==3.10
importlib_metadata==8.5.0
isort==5.13.2
loguru==0.7.3
mccabe==0.7.0
nodeenv==1.9.1
packaging==24.2
platformdirs==4.3.6
pre_commit==4.0.1
pycodestyle==2.12.1
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyflakes==3.2.0
pygal==3.0.5
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.0
zipp==3.21.0
